% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_county_fips.R
\name{add_county_fips}
\alias{add_county_fips}
\title{Make a new county_fips column that is consistent across years}
\usage{
add_county_fips(df, county_vector)
}
\arguments{
\item{df}{cleaned MCOD dataframe with countyoc or countyrs column}

\item{county_vector}{specify if we should use countyoc or countyrs}
}
\value{
same dataframe with new county_fips column
}
\description{
Make a new county_fips column that is consistent across years
}
