% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_asrate_var.R
\name{calc_asrate_var}
\alias{calc_asrate_var}
\title{Calculate age-specific rates and variance}
\usage{
calc_asrate_var(df, new_name, death_col, pop_col = pop)
}
\arguments{
\item{df}{processed MCOD dataframe}

\item{new_name}{bare prefix of the new column names (e.g., opioid)}

\item{death_col}{column of counts for numerator of rate}

\item{pop_col}{column of population for denominator of rate}
}
\value{
dataframe with two new columns
}
\description{
Given a bare (unquoted) column of counts and population, will return the
rate in 100,000 as well as the variance.
}
