% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_age_5.R
\name{categorize_age_5}
\alias{categorize_age_5}
\title{Create a categorical age column from a converted ager27 column}
\usage{
categorize_age_5(age_column)
}
\arguments{
\item{age_column}{age column created from convert_ager27()}
}
\value{
factor
}
\description{
Simply makes more meaningful labels on the age column. Assumes that the age
column was created from the convert_ager27() function. Allows for using
tidyr::complete() even when some age/year/race combinations have no
observations.
}
\examples{
categorize_age_5(seq(0, 70, 5))
}
