% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_icd9_data.R
\name{clean_icd9_data}
\alias{clean_icd9_data}
\title{A wrapper function to perform basic cleaning of ICD-9 dataframes}
\usage{
clean_icd9_data(icd9_df)
}
\arguments{
\item{icd9_df}{an ICD-9 dataframe}
}
\value{
dataframe
}
\description{
Across 1979-1998, MCOD data have minor inconsistencies. This wrapper
function renames nature of injury columns, pads UCOD codes with no
sub-codes, prefixes an E to external cause of injury codes and an N to
nature of injury codes, and trims unnecessary characters and whitespace.
}
