% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_to_fwf_df.R
\name{.dct_to_fwf_df}
\alias{.dct_to_fwf_df}
\title{Downloads the NBER Stata dct file and returns a dictionary with column data}
\source{
http://www.nber.org/data/vital-statistics-mortality-data-multiple-cause-of-death.html
}
\usage{
.dct_to_fwf_df(year)
}
\arguments{
\item{year}{year of the dictionary you want to convert}
}
\value{
dataframe
}
\description{
The NBER has Stata dictionary files for each of the MCOD public use datasets.
This function downloads the dictionary file and converts it to a dataframe
with column name, column start position, column end position, and column
time (in readr::col() format).
}
