% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_1979_pop_data.R
\name{.download_1979_pop_data}
\alias{.download_1979_pop_data}
\title{Download 1979 Population Counts}
\source{
https://www.census.gov/programs-surveys/popest.html
}
\usage{
.download_1979_pop_data()
}
\value{
Dataframe with population counts by age and sex
}
\description{
Download US Census Bureau annual population estimates for 1979 for
each age group and sex. Note that each population estimate file is
a little different and thus must be munged before being combined
into the total pop_est dataframe.
}
