% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_1980s_pop_data.R
\name{.download_1980s_pop_data}
\alias{.download_1980s_pop_data}
\title{Download 1980s Population Counts}
\source{
https://www.census.gov/programs-surveys/popest.html
}
\usage{
.download_1980s_pop_data(raw_folder = "./raw_data", filter_race = TRUE)
}
\arguments{
\item{raw_folder}{location to store downloaded files}

\item{filter_race}{Subset to white, nhw, black, and total (default: TRUE)}
}
\value{
Dataframe with population counts by age and sex
}
\description{
Download US Census Bureau annual population estimates for the 1980s for
each age group and sex. Note that each population estimate file is
a little different and thus must be munged before being combined
into the total pop_est dataframe.
}
