% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_2000s_pop_data.R
\name{.download_2000s_pop_data}
\alias{.download_2000s_pop_data}
\title{Download 2000s Population Counts}
\source{
https://www.census.gov/programs-surveys/popest.html
}
\usage{
.download_2000s_pop_data(filter_race = TRUE)
}
\arguments{
\item{filter_race}{Subset to white, nhw, black, and total (default: TRUE)}
}
\value{
Dataframe with population counts by age and sex
}
\description{
Download US Census Bureau annual population estimates for the 2000s for
each age group and sex. Note that each population estimate file is
a little different and thus must be munged before being combined
into the total pop_est dataframe.
}
