% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_standard_pops.R
\name{.download_standard_pops}
\alias{.download_standard_pops}
\title{Download Standard Populations}
\source{
https://seer.cancer.gov/stdpopulations/
}
\usage{
.download_standard_pops()
}
\value{
Dataframe with SEER standard populations
}
\description{
Download a variety of common standard populations from the SEER website.
Performs minimal manipulation to make age and standard factors human
readable and consistent across standards. Note that one must dplyr::filter()
to a single standard before performing dplyr::left_join() on age_cat.
}
