% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_year.R
\name{.is_icd10}
\alias{.is_icd10}
\title{Is this an ICD-10 data year? (US MCOD adopted ICD-10 from 1999)}
\usage{
.is_icd10(year)
}
\arguments{
\item{year}{a single year}
}
\value{
logical
}
\description{
Is this an ICD-10 data year? (US MCOD adopted ICD-10 from 1999)
}
\keyword{internal}
