% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_mcod_csv.R
\name{download_mcod_csv}
\alias{download_mcod_csv}
\title{Download the multiple cause of death data as a CSV file}
\source{
http://www.nber.org/data/vital-statistics-mortality-data-multiple-cause-of-death.html
}
\usage{
download_mcod_csv(year, download_dir = "./raw_data",
  territories = FALSE)
}
\arguments{
\item{year}{year to download (as integer)}

\item{download_dir}{file path to save download data}

\item{territories}{download US territories}
}
\value{
none
}
\description{
NBER hosts publicly available multiple cause of death data in several
formats (see source). This function downloads the specified year to the
specified folder in the CSV format.
}
