% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_mcod_fwf.R
\name{download_mcod_fwf}
\alias{download_mcod_fwf}
\title{Download the multiple cause of death data as an ASCII file from the CDC}
\source{
ftp://ftp.cdc.gov/pub/Health_Statistics/NCHS/Datasets/DVS/mortality/
}
\usage{
download_mcod_fwf(year, download_dir = "./raw_data")
}
\arguments{
\item{year}{year to download (as integer)}

\item{download_dir}{file path to save downloaded data}
}
\value{
none
}
\description{
The CDC hosts publicly available multiple cause of death data as a fixed-
width text file. This function downloads that file as a zip. Note that the
CDC FTP is very slow--downloading from NBER via the download_mcod_dta() or
download_mcod_csv() functions is strongly suggested.
}
