% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_natality_data.R
\name{download_natality_ascii}
\alias{download_natality_ascii}
\title{Download NCHS Natality (Live Births) Data from the CDC FTP (ASCII)}
\source{
https://www.cdc.gov/nchs/data_access/vitalstatsonline.htm
}
\usage{
download_natality_ascii(year, download_dir = "./raw_data")
}
\arguments{
\item{year}{year to download (as integer)}

\item{download_dir}{file path to save dowhlo data}
}
\value{
none
}
\description{
Fair warning: These files are very large when uncompressed. They range
between 3 and 5 GB uncompressed with a compression ratio of around 90%.
}
