% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_maternal_deaths.R
\name{flag_maternal_deaths}
\alias{flag_maternal_deaths}
\title{Creates a new column called maternal_death with 1 if maternal death}
\usage{
flag_maternal_deaths(
  processed_df,
  year = NULL,
  ucod_only = FALSE,
  keep_cols = FALSE
)
}
\arguments{
\item{processed_df}{MCOD dataframe already processed}

\item{year}{if NULL, will attempt to detect}

\item{ucod_only}{if TRUE, only flag maternal deaths in underlying cause}

\item{keep_cols}{keep intermediate columns}
}
\value{
a new dataframe with a binary maternal_death column
}
\description{
Maternal deaths according to ICD-10 codes.
}
