% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_nonopioid_drug_deaths.R
\name{flag_nonopioid_drug_deaths}
\alias{flag_nonopioid_drug_deaths}
\title{Flag non-opioid drug deaths according to ISW7 rules}
\usage{
flag_nonopioid_drug_deaths(processed_df)
}
\arguments{
\item{processed_df}{processed dataframe}
}
\value{
new dataframe with a nonop_drug_death column
}
\description{
Given an MCOD dataframe **with** drug_death and opioid_death columns
already, will flag non-opioid deaths. Must run flag_opioid_deaths() and
flag_drug_deaths() first.
}
\examples{
df <- data.frame(year = 2019, ucod = "X42", f_records_all = "T509")
df |>
    flag_drug_deaths(year = 2019) |>
    flag_opioid_deaths(year = 2019) |>
    flag_nonopioid_drug_deaths()
}
