% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_other_op_present.R
\name{flag_other_op_present}
\alias{flag_other_op_present}
\title{Creates a column `other_op_present` for deaths with other uspecified opioid}
\usage{
flag_other_op_present(processed_df, year = NULL)
}
\arguments{
\item{processed_df}{MCOD dataframe already processed}

\item{year}{if NULL, will attempt to detect}
}
\value{
a new dataframe with 1 additional column
}
\description{
This function flags all opioid deaths that involved other unspecified opioid
}
