% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_od_intent.R
\name{label_od_intent}
\alias{label_od_intent}
\title{Label intent from underlying cause column for overdose drugs}
\usage{
label_od_intent(processed_df)
}
\arguments{
\item{processed_df}{MCOD dataframe already processed}
}
\value{
dataframe
}
\description{
Label intent from underlying cause column for overdose drugs
}
