% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_suicide_types.R
\name{label_suicide_type}
\alias{label_suicide_type}
\title{Create a new column with labels for suicide type}
\usage{
label_suicide_type(processed_df)
}
\arguments{
\item{processed_df}{processed MCOD dataframe}
}
\value{
dataframe
}
\description{
Create a new column with labels for suicide type
}
\examples{
df <- data.frame(year = 2019, ucod = c("X72", "X68"))
df |>
    flag_suicide_types(year = 2019) |>
    label_suicide_type()
}
