% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narcan.R
\docType{package}
\name{narcan}
\alias{narcan-package}
\alias{narcan}
\title{narcan R package for working with multiple cause of death data}
\description{
R package for working with multiple cause of death data
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/mkiang/narcan}
  \item Report bugs at \url{http://github.com/mkiang/narcan/issues}
}

}
\author{
\strong{Maintainer}: Mathew Kiang \email{mathew.kiang@gmail.com}

Authors:
\itemize{
  \item Monica Alexander \email{monicaalexander@berkeley.edu}
}

}
