% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_3char_codes.R
\name{pad_3char_codes}
\alias{pad_3char_codes}
\title{Pad ICD-9 codes that do not have a sub-code (i.e., 3-character codes)}
\usage{
pad_3char_codes(icd9_ucod)
}
\arguments{
\item{icd9_ucod}{The ucod column from an ICD-9 dataframe}
}
\value{
vector
}
\description{
Some of the ICD-9 codes only contain 3 characters (i.e., they do not contain
subcodes). This function just takes the UCOD column and converts all of
these 3 character codes into 4 by padding a zero at the end.
}
\examples{
pad_3char_codes(c("400", "4043", "304", "5062"))
}
