% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_ni_flag.R
\name{rename_ni_flag}
\alias{rename_ni_flag}
\title{Rename nature of injury columns for consistency}
\usage{
rename_ni_flag(icd9_df)
}
\arguments{
\item{icd9_df}{an ICD-9 dataframe}
}
\value{
dataframe
}
\description{
Some of the NBER ICD-9 MCOD files use rniflag_ as a column name while
others use rnifla_. This function makes the names consistent (to rnifla_).
}
\examples{
df <- data.frame(rniflag_1 = c(0, 1), record_1 = c("E850", "9650"))
rename_ni_flag(df)
}
