% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-std_pops.R
\docType{data}
\name{std_pops}
\alias{std_pops}
\title{Dataframe of common standard populations in 18 age categories}
\format{A data frame with 234 rows and 5 columns
\describe{
  \item{age_cat}{factor, age group}
  \item{standard_cat}{factor, description of standard population}
  \item{pop_std}{count, population in that age group for that standard}
  \item{standard}{code, character code for that standard population}
  \item{age}{starting age of that age group}
}}
\source{
\url{https://seer.cancer.gov/stdpopulations/stdpop.18ages.txt}
}
\usage{
std_pops
}
\description{
A data set containing a variety of common standard populations including:
2000 US standard million, 2000 US standard population, 1960 world standard
million, WHO standard million. Created by using the download_standard_pops()
function.
}
\keyword{datasets}
