% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_5char_record.R
\name{trim_5char_record}
\alias{trim_5char_record}
\title{Trim ICD-9 record columns that include the nature of injury flag}
\usage{
trim_5char_record(record_col)
}
\arguments{
\item{record_col}{The record column from an ICD-9 dataframe}
}
\value{
vector
}
\description{
For some years, the ICD-9 record columns are 5 character codes with the last
character representing the nature of injury (N) flag. Just trim off these
columns and use the appropriate rnifla_ column for the N flag.
}
\examples{
trim_5char_record(c("400 1", "40000", "400", "400 "))
}
