% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_records.R
\name{unite_records}
\alias{unite_records}
\title{Unite the 20 record columns from ICD-9 dataframe into a single column}
\usage{
unite_records(icd9_df)
}
\arguments{
\item{icd9_df}{an ICD-9 dataframe}
}
\value{
dataframe
}
\description{
ICD-9 dataframes contain 20 record columns (contributory causes) along with
20 flag columns that denote if the contributory cause is an N code or an
E code. This function collapses the 20 columns into a single column with
all codes (and their appropriate prefix) to make regex'ing easier.
}
