% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hspanicr_column.R
\name{add_hspanicr_column}
\alias{add_hspanicr_column}
\title{Add an NA hspanicr column if one doesn't exist}
\usage{
add_hspanicr_column(icd_df)
}
\arguments{
\item{icd_df}{an MCOD dataframe)}
}
\value{
dataframe
}
\description{
Hispanic origin was not recorded until 1989. In order to keep all dataframes
comformable, add an NA column named hspanicr if one does not exist.
}
