% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_populations.R
\name{add_pop_counts}
\alias{add_pop_counts}
\title{Given a dataframe with year, age, sex, and race, returns population estimate}
\usage{
add_pop_counts(df, by_vars = c("year", "age", "sex", "race"))
}
\arguments{
\item{df}{MCOD dataframe}

\item{by_vars}{variables to match on}
}
\value{
dataframe
}
\description{
Uses the internal narcan::pop_est dataset to return yearly population
estimates by age, sex, and race. See narcan::pop_est for columns and
possible values of each matching variable
}
