% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_populations.R
\name{add_std_pop}
\alias{add_std_pop}
\title{Given a dataframe with age, returns a standard population}
\usage{
add_std_pop(df, std_cat = "s204", by_vars = "age")
}
\arguments{
\item{df}{dataframe with age column (in 5-year bins)}

\item{std_cat}{standard population to use (default: US 2000 standard pop)}

\item{by_vars}{variables to merge on}
}
\value{
dataframe
}
\description{
Returns the US 2000 population by default, by any population from the
narcan::std_pop$std_cat column is valid.
}
