% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stdrate_var.R
\name{calc_stdrate_var}
\alias{calc_stdrate_var}
\title{Calculate age-standardized rates and variance}
\usage{
calc_stdrate_var(df, asrate_col, asvar_col, ..., weight_col = unit_w)
}
\arguments{
\item{df}{processed MCOD dataframe}

\item{asrate_col}{age-specific rate column}

\item{asvar_col}{variance of the age-specific rate}

\item{...}{grouping variables (in addition to year and race)}

\item{weight_col}{column of (unit) weights}
}
\value{
dataframe with two new columns
}
\description{
Given a bare (unquoted) column of age-specific rates, variance, and weights,
will return the age-standardized rate and variance.
}
