% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ager27u1.R
\name{convert_ager27u1}
\alias{convert_ager27u1}
\title{Converts the age27 variable in MCOD data to under-1, 1-4, then 5-year groups}
\usage{
convert_ager27u1(icd_df, remove_age27 = TRUE)
}
\arguments{
\item{icd_df}{an MCOD dataframe with age27 as a column}

\item{remove_age27}{once a new column is created, remove the old age27}
}
\value{
dataframe
}
\description{
Converts the age27 variable in MCOD data to under-1, 1-4, then 5-year groups
}
