% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_mcod_fwf_dicts.R
\name{.download_mcod_fwf_dicts}
\alias{.download_mcod_fwf_dicts}
\title{Download Fixed Width Format Dictionaries}
\source{
https://www.nber.org/research/data/mortality-data-vital-statistics-nchs-multiple-cause-death-data
}
\usage{
.download_mcod_fwf_dicts()
}
\value{
Dataframe with MCOD fixed width format data
}
\description{
Download the fixed width format dictionaries from NBER for 1979 to 2018,
then add 2020 manually (2020 added occupation) and duplicate 2018 to 2019.
}
