% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_year.R
\name{.extract_year}
\alias{.extract_year}
\title{Helper function that extracts year from a dataframe and raises error
if more than one.}
\usage{
.extract_year(df)
}
\arguments{
\item{df}{dataframe to extract year from}
}
\value{
year as integer
}
\description{
Helper function that extracts year from a dataframe and raises error
if more than one.
}
