% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_restricted_data.R
\name{.import_restricted_data}
\alias{.import_restricted_data}
\title{Wrapper for importing restricted MCOD data}
\usage{
.import_restricted_data(file, year, fix_states = TRUE)
}
\arguments{
\item{file}{path to restricted MCOD plaintext file}

\item{year}{year of MCOD data}

\item{fix_states}{recode state abbreviations to their FIPS code}
}
\value{
dataframe
}
\description{
Restricted MCOD data contains geographical location (for years after 2004),
that the public-use files do not contain. Further, restricted files come as
plaintext, fixed-width files. This helper function simply imports these
text files with known dictionaries.
}
