% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_mcod_dta.R
\name{download_mcod_dta}
\alias{download_mcod_dta}
\title{Download the multiple cause of death data as a DTA file}
\source{
http://www.nber.org/data/vital-statistics-mortality-data-multiple-cause-of-death.html
}
\usage{
download_mcod_dta(year, download_dir = "./raw_data")
}
\arguments{
\item{year}{year to download (as integer)}

\item{download_dir}{file path to save downlaoded data}
}
\value{
none
}
\description{
Download the multiple cause of death data as a DTA file
}
