% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_natality_ascii.R
\name{download_natality_ascii}
\alias{download_natality_ascii}
\title{Download NCHS Natality (Live Births) Data from the CDC FTP (ASCII)}
\source{
https://www.cdc.gov/nchs/data_access/vitalstatsonline.htm
}
\usage{
download_natality_ascii(year, download_dir = "./raw_data", return_path = FALSE)
}
\arguments{
\item{year}{year to download (as integer)}

\item{download_dir}{file path to save downloaded data}

\item{return_path}{return the path of the file that was downloaded}
}
\value{
none
}
\description{
Fair warning: These files are very large when uncompressed. They range
between 3 and 5 GB uncompressed with a compression ratio of around 90%.
Also, it appears the CDC rate limits downloads so I do not export this
function. To use it, you'll need to use the triple colon:
`narcan:::download_natality_ascii()`.
}
\details{
Further, the unzip() function in R will not unzip files that are larger
than 4GB so this needs to be unzipped using a system call or external
program.
}
