% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_drug_deaths.R
\name{flag_drug_deaths}
\alias{flag_drug_deaths}
\title{Flag drug deaths according to ISW7 rules}
\usage{
flag_drug_deaths(processed_df, year = NULL, keep_cols = FALSE)
}
\arguments{
\item{processed_df}{processed dataframe}

\item{year}{if NULL, will attempt to detect}

\item{keep_cols}{keep intermediate columns}
}
\value{
new dataframe with a drug_death column
}
\description{
Given an MCOD dataframe, will apply ISW7 rules to flag drug deaths for
both ICD9 and ICD10 codes.
}
