% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_methadone_present.R
\name{flag_methadone_present}
\alias{flag_methadone_present}
\title{Creates a column `methadone_present` if opioid death involved methadone}
\usage{
flag_methadone_present(processed_df, year = NULL)
}
\arguments{
\item{processed_df}{MCOD dataframe already processed}

\item{year}{if NULL, will attempt to detect}
}
\value{
a new dataframe with 1 additional column
}
\description{
This function flags all opioid deaths that involved methadone.
}
