% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_nonheroin.R
\name{flag_nonheroin}
\alias{flag_nonheroin}
\title{Flag all opioid deaths that were not from heroin}
\usage{
flag_nonheroin(processed_df)
}
\arguments{
\item{processed_df}{MCOD dataframe with flag_opioid_types() columns}
}
\value{
dataframe
}
\description{
NOTE: assumes flag_opioid_types() has already been run.
}
