% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_od_intent.R
\name{flag_od_intent}
\alias{flag_od_intent}
\title{Flag overdose deaths by their UCOD intent code}
\usage{
flag_od_intent(processed_df, year = NULL)
}
\arguments{
\item{processed_df}{MCOD dataframe already processed}

\item{year}{if NULL, will attempt to extract}
}
\value{
dataframe
}
\description{
Flag overdose deaths by their UCOD intent code
}
\examples{
df <- data.frame(
    year = 2019,
    ucod = c("X42", "X62"),
    f_records_all = c("T400", "T400")
)
df |>
    flag_drug_deaths(year = 2019) |>
    flag_od_intent(year = 2019)
}
