% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_opioid_contributed.R
\name{flag_opioid_contributed}
\alias{flag_opioid_contributed}
\title{Flag non-opioid deaths that involved opioids}
\usage{
flag_opioid_contributed(processed_df, year = NULL)
}
\arguments{
\item{processed_df}{processed dataframe}

\item{year}{if NULL, will attempt to detect}
}
\value{
new dataframe with an opioid_contributed column
}
\description{
Given an MCOD dataframe, will search through contributory causes for
opioid related codes, but will only flag rows for which the underlying
cause is **NOT** an opioid-related death.
}
\details{
NOTE: This function really doesn't make sense for ICD-9 years. Use with
caution.
}
