% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_opioid_deaths.R
\name{flag_opioid_deaths}
\alias{flag_opioid_deaths}
\title{Flag opioid deaths according to ISW7 rules}
\usage{
flag_opioid_deaths(processed_df, year = NULL)
}
\arguments{
\item{processed_df}{processed dataframe}

\item{year}{if NULL, will attempt to detect}
}
\value{
new dataframe with an opioid_death column
}
\description{
Given an MCOD dataframe, will apply ISW7 rules to flag opioid deaths for
both ICD9 and ICD10 codes.
}
