% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_opioid_types.R
\name{flag_opioid_types}
\alias{flag_opioid_types}
\title{Take a processed MCOD dataframe and create indicators for opioid types}
\usage{
flag_opioid_types(processed_df, year = NULL)
}
\arguments{
\item{processed_df}{MCOD dataframe already processed}

\item{year}{if NULL, will attempt to detect}
}
\value{
a new dataframe with 9 additional columns
}
\description{
Creates 9 indicators for all opioid deaths. 7 for type of opioid (opium,
heroin, natural, methadone, synethtic, other, unknown), 1 column for
the number of opioids, and 1 column to indicate presence of more than one opioid.
}
