% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_opium_present.R
\name{flag_opium_present}
\alias{flag_opium_present}
\title{Creates a column called `opium_present` if opioid death involved opium}
\usage{
flag_opium_present(processed_df, year = NULL, missing_val = 0)
}
\arguments{
\item{processed_df}{MCOD dataframe already processed}

\item{year}{if NULL, will attempt to detect}

\item{missing_val}{value to indicate missing (i.e., code did not exist)}
}
\value{
a new dataframe with 1 additional column
}
\description{
Note that ICD9 years did not include an opium code. We code it as 0 by
default, but can be coded however specified in missing_val parameter.
This function flags all opioid deaths that involved opium.
}
