% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_suicide_deaths.R
\name{flag_suicide_deaths}
\alias{flag_suicide_deaths}
\title{Flag suicide deaths (no accidental poisoning)}
\usage{
flag_suicide_deaths(df)
}
\arguments{
\item{df}{processed MCOD dataframe}
}
\value{
new dataframe
}
\description{
TODO: Make this also work with ICD-9
}
