% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_suicide_types.R
\name{flag_suicide_fall}
\alias{flag_suicide_fall}
\title{Flag suicide by fall}
\usage{
flag_suicide_fall(df)
}
\arguments{
\item{df}{a processed MCOD dataframe}
}
\value{
dataframe
}
\description{
TODO: MAKE WORK WITH ICD-9
}
