% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_suicide_types.R
\name{flag_suicide_other}
\alias{flag_suicide_other}
\title{Flag suicide by other (not poison, fall, firearm, suffocation)}
\usage{
flag_suicide_other(df)
}
\arguments{
\item{df}{a processed MCOD dataframe}
}
\value{
dataframe
}
\description{
ICD-10 codes: U03, X71, X75-X79, X81-X84, Y870
}
\details{
TODO: MAKE WORK WITH ICD-9
}
