% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_suicide_types.R
\name{flag_suicide_types}
\alias{flag_suicide_types}
\title{Flag suicide five types: firearm, poisoning, fall, suffocation, or other}
\usage{
flag_suicide_types(df)
}
\arguments{
\item{df}{processed MCOD dataframe}
}
\value{
new dataframe
}
\description{
TODO: Make this also work with ICD-9
}
