% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_suicide_types.R
\name{label_suicide_type}
\alias{label_suicide_type}
\title{Create a new column with labels for suicide type}
\usage{
label_suicide_type(processed_df)
}
\arguments{
\item{processed_df}{processe MCOD dataframe}
}
\value{
dataframe
}
\description{
Create a new column with labels for suicide type
}
