% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mcod_fwf_dicts.R
\docType{data}
\name{mcod_fwf_dicts}
\alias{mcod_fwf_dicts}
\title{Dataframe of fixed width format information for MCOD files}
\format{
A data frame with 5 columns
\describe{
  \item{name}{string, variable name}
  \item{type}{string, variable type (i.e., numeric, character, etc.)}
  \item{start}{integer, starting position of this column}
  \item{end}{integer, ending position of this column}
  \item{year}{integer, year of this dictionary}
}
}
\source{
\url{https://www.cdc.gov/nchs/nvss/mortality_public_use_data.htm}
}
\usage{
mcod_fwf_dicts
}
\description{
Dataframe of fixed width format information for MCOD files
}
\keyword{datasets}
