% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkdir_p.R
\name{mkdir_p}
\alias{mkdir_p}
\title{Wrapper to make a directory (and subdirectories) if necessary}
\usage{
mkdir_p(path)
}
\arguments{
\item{path}{Location of folder to be made}
}
\value{
None
}
\description{
Mimics `mkdir -p` by making a directory and all subdirectories and
suppressing error messages if folder already exists.
}
\examples{
mkdir_p('./test_directory')
}
