% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narcan.R
\docType{package}
\name{narcan}
\alias{narcan}
\alias{narcan-package}
\title{narcan R package for working with multiple cause of death data}
\description{
narcan R package for working with multiple cause of death data
}
