% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_e_to_ucod.R
\name{prefix_e_to_ucod}
\alias{prefix_e_to_ucod}
\title{Add the external cause flag (E) to appropriate ICD-9 UCOD codes}
\usage{
prefix_e_to_ucod(icd9_ucod)
}
\arguments{
\item{icd9_ucod}{The ucod column from an ICD-9 dataframe}
}
\value{
vector
}
\description{
All UCOD ICD-9 codes between 800 and 999 are external cause of injury (E)
codes. We append the E to them to make regexing across UCOD and record
columns consistent.
}
\examples{
prefix_e_to_ucod(c(7951, 8001, 9992, 6000, 4000))
}
