% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_abbrev_to_fips.R
\name{state_abbrev_to_fips}
\alias{state_abbrev_to_fips}
\title{Replace state abbreviations with their corresponding FIPS code}
\usage{
state_abbrev_to_fips(column)
}
\arguments{
\item{column}{a vector of strings with state abbreviations}
}
\value{
a new vector with state FIPS
}
\description{
Replace state abbreviations with their corresponding FIPS code
}
