% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_residents.R
\name{subset_residents}
\alias{subset_residents}
\title{Subset to US residents}
\usage{
subset_residents(df, drop_col = TRUE)
}
\arguments{
\item{df}{an MCOD dataframe}

\item{drop_col}{drop the `restatus` column after subsetting (default: TRUE)}
}
\value{
dataframe
}
\description{
Subset to US residents
}
