% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_binary_columns.R
\name{summarize_binary_columns}
\alias{summarize_binary_columns}
\title{Summarizes all flagged (e.g., 0/1) MCOD columns}
\usage{
summarize_binary_columns(df, ...)
}
\arguments{
\item{df}{a dataframe with binary flag columns to indicate type of death}

\item{...}{grouping variables (in addition to year and age)}
}
\value{
dataframe
}
\description{
To use this, you must remove all non-grouping, non-binary variables.
}
