% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_records.R
\name{unite_records}
\alias{unite_records}
\title{Unite the 20 record columns from MCOD dataframe into a single column}
\usage{
unite_records(icd_df, year = NULL)
}
\arguments{
\item{icd_df}{an ICD dataframe}

\item{year}{the year of this dataframe -- if NULL, will attempt to detect}
}
\value{
dataframe
}
\description{
This function collapses the 20 contributory cause columns into a single
column for easier regex'ing. ICD-9 dataframes will also get appropriate
prefixes before collapsing.
}
